+++ internal/app/postgres/country.go
@@ -63,7 +63,7 @@ func (p *CountryRepository) Update(c *app.Country) error {
 
 // Retrieve an existing domain Country struct from database, errors that occur while performing the
 // operation are returned.
-func (p *CountryRepository) GetById(id int) (*app.Country, error) {
+func (p *CountryRepository) GetById(id int64) (*app.Country, error) {
 	query := `SELECT * from sportmonks_country where id = $1`
 	row := p.connection.QueryRow(query, id)
